% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.r
\name{compute}
\alias{compute}
\title{Deprecated function}
\usage{
compute(x, covariate, rep = 1)
}
\arguments{
\item{x}{an object of class \code{nn}.}

\item{covariate}{a dataframe or matrix containing the variables that had
been used to train the neural network.}

\item{rep}{an integer indicating the neural network's repetition which
should be used.}
}
\value{
\code{compute} returns a list containing the following components:
\item{neurons}{a list of the neurons' output for each layer of the neural
network.} \item{net.result}{a matrix containing the overall result of the
neural network.}
}
\description{
The function \code{compute} is deprecated. Please refer to the new function \code{\link{predict.nn}}.
}
